Moved preprocessor resolution into seperate file
