+++ plugin/src/types.d.ts
@@ -1,0 +1,6 @@
+import type { preprocess, parse } from 'svelte/compiler'
+import type {Plugin} from 'vite'
+ 
+type Config = Parameters<Plugin["configResolved"]>[0];
+type PreprocessorGroup = Parameters<typeof preprocess>[1];
+type Ast = ReturnType<typeof parse>
